Global P2P Checkpoint
- Added Global Sync P2P checkpoint - (CHECKPOINT_DISTRIBUTED_MODE=2);
When a miner solves a block it is immediately broadcast globally; All
stake miners try to accept the block and start hashing a new block
* To test this:  The log should show Global Sync Checkpoint broadcast
xxx (xxx is last blockhash) - Receiving nodes should log: Received
Global Checkpoint xxx; Should start hashing a new block;
* If this is used as an attack the receiving nodes will ignore the
checkpoint after one minute - Or if the block is not received within one
minute (In this condition the log states: Never received global
solution; purging; )
- Added dynamic subsidy mint limiter for small mint: This allows small
block mint subsidies when diff is low
Mint Limiter schedule:
if (PORDiff >= 0  && PORDiff < .5) return .0001;
if (PORDiff >= .5 && PORDiff < 1)  return .25;
if (PORDiff >= 1  && PORDiff < 5)  return 1;
if (PORDiff >= 5  && PORDiff < 10) return 5;
if (PORDiff >= 10 && PORDiff < 50) return 7;
if (PORDiff >= 50) return 10;
Note: The client will still stake a small mint 10% of the time
regardless of the mint limiter to ensure users with very small balances
can stake.  This rule is in effect 90% of the time and will not cause a
block rejection on the accept side.