3.3.0.1
Gridcoin Research 3.3.0.1/14.1
Leisure Upgrade
MSI:
http://download.gridcoin.us/download/downloadstake/GridcoinResearch.msi

- CPID connection support on linux
- Allow newbie to stake first block easily
- Partial support for cpidv2 :
(Last resort to disable cpidv2 temporarily: disablecpidv2=true) : NOTE:
This was tested on linux, this should not be required.
- Added accuracy to block.
- Running Boinc Remotely (or without boinc installed): From the node
with boinc and or BAM installed do this once: execute genboinckey
Copy the key into the config of the node without boinc installed:
Example: boinckey=xxx;
We will eventually work on removing the necessity to go through this
step when we work on ease-of-use features.
- Fixed issue : cannot sync from 0
- Diff reduction feature and improved mintlimiter rules