3.3.1.0
Gridcoin Research 3.3.1.0/15.0
Mandatory Upgrade
MSI:
http://download.gridcoin.us/download/downloadstake/GridcoinResearch.msi

- Optional feature for windows users : Suppress Windows Error Reporting
Dialog : to set this key execute this command from the rpc (This means
that
when an app crashes on the node, you will not receive the Windows Error
Reporting Dialog - Would you like to check the internet for solutions or
Debug this program?  The program will just end.)
execute DISABLE_WINDOWS_ERROR_REPORTING
(Note: Crash dumps will still be created and stored in the standard
location if you set this key).
(This feature was requested 90 days ago in testnet - its not really
needed with a stable client, but may be useful when we go back to
testnet).
- Added a generic interface to the win32 side to allow consolidation of
the commands and for extensibility.
- CPIDV2 integrated on all blocks (still need to remove CPIDV1 in 5
days)
- Check up to 3 times with netsoft before rejecting the block in the
second place in Gridcoin
- More anti-orphan adjustments