3.3.9.8
Gridcoin Research 3.3.9.8/23.1
Leisure Upgrade

- Removed one line of repetetive logging
- Added generic data storage system to block chain
- Added RPC commands : 'execute addkey', 'execute memorizekeys',
'execute listdata'
Real World Examples:
a. Storing a price for a smart contract (permanently in the chain):
(Note that anyone can add generic data to the chain for 1grc - the
proceeds go to the foundation node - each add charges a 1 grc fee)
(If the data will be used by the coin for business purposes it will be
done in a different manner - IE not through the public RPC - with the
private key pair for security).
Storing a price in the chain:
execute addkey add price grc .0000046
(Note that this price can be updated with another command):
execute addkey add price grc .0000045
However, the last price in the highest block always contains the current
state of the data.
You may delete the value like this:
execute addkey delete price grc .0000045
The syntax for the command is:
execute addkey [add or delete] [key type] [key name] [key value]
(Data is stored permanently in the chain and is never physically
erased).

Note:  In this release you must manually execute a command to load the
data values into the coin before you can list them in the rpc (in later
versions this will be done automatically):

execute memorizekeys
To see the data for a specific data type:
execute listdata [key type]
So, to see the price above since it belongs to the "price" key type:
execute memorizekeys
execute listdata price

Note: You can store data in any key type

To see current whitelisted projects (Only dummy keys exist):
execute listdata project

...

Now a simple all in one example for those confused:
execute addkey add price grc .0000046
(wait for block to be accepted in tx list otherwise price is not
changed)
execute memorizekeys
execute listdata price
(This last command shows the price you changed... etc.)

For RTM, the projects can be whitelisted or blacklisted using the
following commands.  Note that an admin must have the boinc project
admin key stored in the config file to modify system data:
execute addkey add project milkyway 1
(This adds milkyway to the whitelist; we will only need to modify the
non case sensitive boinc project name - we can do away with the http://
URL)
execute addkey delete project milkyway 1
(This removes milkyway from the whitelist).

Note that after a key is added you have to wait until the block is
accepted by the network before you can do an "execute memorizekeys" in
order for the values to be available.

- UPNP: Marix fixed some issues with UPNP and QR code generation.
- Dr Grid : ToolTips: we had a contribution to make the hover over
tooltips on the overview look better, and change the font color for
tooltips.
- Contribution: Timestamp for CSV mangitude report (list magnitudecsv)
- Marix: Change detectplatform and remove some files in Build process
- IceD: Change some Boost suffixes
- Marix : Change to Leder's upgrader
- Caraka:  Improved toolbar icon.