3.4.0.6
Gridcoin Research 3.4.0.6/24.0
Leisure Upgrade

Note:  Due to the complexity of DPOR, the feature is broken into phases.
In Phase I, we will need to reconcile the DPOR rac to match prod, thus
it would be prudent for us to test this version
together in prod (we are running DPOR side-by-side the legacy system to
maintain complete network compatibility).  I think testnet will be too
small and error prone to complete this exercise.  Instructions below:

- Incremented minor version to force upgrade in testnet to resolve Quez
& Robs problem (this will NOT force a mandatory upgrade).  I believe the
issue is fixed (strictly testnet).
- Added smart contract expiration date
- Added ability to query Project URL RAC by team, and World Community
Grid Team RAC.  Added DNS to the coin (to access Boinc URLs by DNS
address).
- Added ability to tally magnitude by on-file smart contracts
- Added DPOR magnitude to list magnitudecsv (last column).  Lets use
this to compare the DPOR Magnitude to the Magnitude.  We will need to
reconcile this in prod and get the entire network RAC
to match before we can rely on the smart contracts.
- Added DPOR magnitude to list rsa.
- Added a service to allow a DPOR node to submit a smart contract for an
unsynchronized project once per day.  Note you will be charged .25grc
when this is done.  We can work on reimbursement in the POR reward soon.
- Added list opencontracts to run the Open Contracts report.  This
report shows the contract name, expiration date, and contents.  We will
need to run this and find boinc projects with missing CPIDs,
when we know network credits already exist in list network for the
project.
Most likely, the URL given to RTM to whitelist the project was wrong.
Note, CM supplied a list of URLs with TeamIDs in the google workbook:

https://docs.google.com/spreadsheets/d/19RFsusf1BSoi4R28d-ul3bOtYGJXafUnXKKHt-gPC2U/pubhtml
(CM can you please make it r/w so we can edit it?)  A couple of the team
IDs are incorrect and that is the reason some contracts are empty in
this report.
Lets add the missing URLs (I already added Bitcoin Utopia) and correct
the team IDs on the projects in list opencontracts with empty cpid list.
RTM can then add that row to the whitelist.
For any contract that is missing, navigate to this URL:
http://boinc.bakerlab.org/team_lookup.php?team_name=gridcoin&format=xml
But replace the boinc.bakerlab.org with the URL of the project.  Find
the XML "id" for the project, and submit it to the forum so RTM can
re-whitelist that project.
If anyone needs to know what projects are whitelisted in DPOR as
compared to legacy, 'list network' shows legacy whitelisted projects,
while "execute listdata project" shows the DPOR whitelisted projects.
Any project missing from that list will need added with our correct Team
ID.
- If your DPOR magnitude does not equal your legacy magnitude, post your
cpid on the forum and we will run the netsoft user report and find which
project needs whitelisted.
Im working on making the report available as an RPC command with a CPID
parameter, and then we can compare Netsoft RAC vs DPOR rac line by line
(didnt quite finish it yet).
- Once we fix the missing projects the DPOR magnitude should be very
close to the legacy magnitude for everyone in list magnitudecsv.  Once
that occurs we can move to Phase II to start enabling DPOS.