3.4.0.8
Gridcoin Research 3.4.0.8/24.2
Leisure Upgrade

- Removed fees charged for generating a DPOR contract once per day
- Ensured we are honoring wallet locks
- Restricted contract creation to POR nodes only
- Limit new contracts to one per 12 hours per node to avoid nuisance
contract creation
- Added execute rac CPID (this command will pull in the netsoft rac
report, the DPOR rac report, and show the RAC side by side to find where
the issues are in DPOR)
This report will also show the Netsoft XML to help reconcile the
difference between the DPOR rac & Netsoft rac
- Removed a Huge amount of logging

The primary reason I am releasing this version is 3.4.0.6/7 had a huge
amount of logging in it, so please upgrade if you want to cut down the
log size.
We can primarily test this in testnet.  But I will need some help in
reconciling the DPOR rac per cpid.
You can type "execute rac YourCpid" and you should receive a report that
shows the differences in project rac comparing DPOR to Netsoft.  That
would be a good place for us to start.