3.4.5.0-Leisure
Gridcoin Research 3.4.5.0/MSI=26.7
Leisure Upgrade

- Added 'tally' field to getinfo and to overview: shows the minimum
block height the coin has memorized while creating everyones RSA account
- Added QuorumHashingAlgorithm to replace md5 for neural network
communication.  This allows NN nodes to agree on everyones magnitude
more quickly.
(The previous NN hash is left in for legacy support until the next
mandatory, so superblocks can now stake using the old system or the new
system).
We will have to have a lot of nodes to upgrade to see the neural report
content decrease to a much higher majority.  We will need a majority of
windows nodes to upgrade for the new system to work.
Next mandatory, we can remove legacy hash.
- Added Pending superblock to 'execute neuralreport'.  This shows the
block number of an accepted pending superblock that has not yet been
loaded by the coin.  The coin loads the superblock as a team 15 blocks
after the block is staked successfully.
- Added contract beacon count and contract hash to contract details for
debugging - tested QHA successfully
- Only add neural hashes to blocks if superblock is old and NN is live
and working - to clean up the neural report
- Removed more unecessary logging and fixed problem where large blocks
were not being forwarded to other nodes and becoming orphans
- Made list explainmagnitude connect to the neural network and pull down
the actual mag report for the user for consistency