3.4.7.7/MinorVersion=294 - Leisure - Testnet
Gridcoin Research 3.4.7.7/294/MSI=28.5
Leisure Upgrade
(Mandatory for TestNet)

- Fixed mintlimiter rule in TestNet keeping nodes from staking
- Added MiningErrors8 to getmininginfo and to overview.  If you are
probing coin age for more than 10 minutes, this field will show what is
going on in the stakeminer.
If it says, Kernel Found 0: that means coin age is not mature enough.
After approx 5 mins, Kernel Found should increase to a portion of the
coins in the Stake bucket.  Once that happens the node will
stake as soon as it passes the mintlimiter threshhold for the diff
level.  Tested in testnet successfully (we should see more staking).
- Added Lifetime Average Magnitude to list rsa.  This will be valuable
to see the long term payments per day per magnitude in the future.