3.5.3.1-leisure
Gridcoin Research 3.5.3.1/MSI=37.2

- Removed netsoft from core for CPID detection and investor status.
- Upgraded Neural Network to pull RAC from the project sites.
Note that the first time you sync, the system will sync slowly while it
pulls the project and team data.
Successive syncs in mid-day will be fast.
The NN will attempt to do a full refresh once every 24 hrs in the
background OR when the superblock is exactly older than 86400 as a team.
- Fixed Investor receiving CPID issue.