3.5.5.8-leisure
Gridcoin Research 3.5.5.8/MSI=39.6
Leisure Upgrade

- Modified GRCRestarter so that downloadblocks uses the new 960gb
anti-ddos host (and modified height to be correct during download)
- Fixed bug in Foundation Expense/Campaign Add to not throw a poll error
while adding an expense or campaign
- Added a new RPC command to allow data to be stored in burn
transactions, and modified the RPC to report on the data (or custom
beacons) in listreceivedbyaddress:
a. To burn gridcoins with a custom data payload, execute the following:
execute burn GRCBurnAddress GRCBurnAmount BurnCategory
This_Is_MyBurnMessage
Note that you create your own Category (for example if your project is
called Asteroids, maybe use Asteroids for the category).
To report on the burns, you may find the TxIds and BurnCategory and Burn
Messages with this command:
listreceivedbyaddress
Note: The output will be JSON and the Key and Data values will only be
in the JSON when they are populated with data.