Fix compilation of Thread priority handling on Linux
Note that on Linux normal users won't actually get a higher-than-normal-
priority effect unless they run as root.

All the code in this area is lacking even basic error checks. This issue
is not touched by this commit, but should be looked at, as this might
cause quite some unexpected behaviour if things go wrong and the code
just continues without saying anything.