3.5.8.0-leisure
Gridcoin Research 3.5.8.0/MSI=41.5
Mandatory Upgrade for TESTNET
Leisure Upgrade for PROD

- At block 196550 in TestNet we require POR blocks to be signed (with
the CPID private key)
- Added 'list magnitude CPID*' where CPID* is Part of your cpid, for
example if your cpid starts with 784, list magnitude 784 will show the
magnitude for that CPID as of the last superblock.
This should cut down confusion if your magnitude is showing as 0 or if
you wonder what value your next POR will be assessed at.
- Fixed problem where the chain rolls back to the next mandatory block
cutover in testnet.
- Added patch to store 0 or 1 in nContract field (please test after
block 196550)