Don't set or clear the stylesheet if nothing has changed.
It seems that on KDE this will trigger an endless loop of focus events which trigger new stylesheet assignments. This bug is only verified to happen on environments which run a KDE theme such as Breeze.
The bitcoin wallet seems to not be affected by this. At least the binary release. No tests have been done on the source release.