Remove four unused variables from CBlockIndex to save RAM.
The CBlockIndex instances are by far the biggest offenders when it comes to
RAM consumption in the Gridcoin wallet. By removing these unused variables
the wallet uses 92MB less RAM and it slows down the pace at which it's
growing.

The "sReserved" flag has been removed from memory but its spot in the
(de)serialization is maintained. Instead of (de)serializing a member
variable we just operate on a local one.