Move FindBlockByHeight functions to a separate class.
This originated from trying to fix a crash bug on testnet. Due to the way the
cache worked it was very tricky to unit test the functionality I it was
extracted to its own class. Every caller can now carry its own cache and there
are fewer static states hanging around. All the various block finding
implementations have been merged into one. It should also be easy to adapt the
finder to parallell find if needed in the future.