Remove Gridcoin address from blocks to save memory.
The intention was to map an address to a CPID but it was never used fully.
This address was used in the CSV export. In order to preserve the
functionality of existing CSV parsers the address field has been kept but
is now always empty. Parsers should prepare for removal of this field in
the future.