Remove coin tracking.
This is a centralized feature which relies on an Azure SQL server being available. Since we are straying away from centralization this needs to go.
It is a neat feature but other cryptos do fine without it. If there's a real need for it we can reimplement it in a decentralized fashion.