Change V8 block switch from 1010000 to 1010001.
The previous fix made it so V8 could be created on 1010001. However, that
also means that the current version and the one after it will behave
differently when the first V8 block arrives. Non-upgraded clients will
reject it while updated ones will accept it. With this fix both versions
will behave the same, but only the upgraded wallets are able to create the
very first V8 block.