Apply predicate on all appcache items, not just the first.
The implementation got a bit more convoluted than I would have liked, but it
still allows us to use ranged based for loops. Since everything moved to the
header there is currently no need for the cpp. Once the appcache accesses
are fully encapsulated we can bring the cpp back.

This closes #682.