Interim new logo and splash screen
This is not to undermine @skcin's great work on pull #517 High dpi improvements. I have replaced the icon, icons, and splash3 png files in order to get our new branding out to the public while using the wallet. The app logo and startup splash screen is the only thing changed. So use if desired.