Change data structure to a map of maps.
This eliminates the need to string manipulation, and it makes
traversing an entire section much easier. We no longer have
to use AppCacheFilter but can iterate directly on the section.

Additionally the cache and the timestamp maps have been merged
into one. Instead of value+timestamp in two different structs, they
both now reside in an AppCacheEntry inside the application cache
data structure.