Make auto upgrades opt in.
Change wallet checks to be time based instead of height based. This removes
the need for checking if projects are initialized and it takes block sync
states out of the equation.

All (Windows) wallet will now check for updates once per day. If the user
has enabled auto upgrades through the "autoupgrade" config option the
upgrade will be downloaded and applied.