Avoid creating threads on each RPC request.
Since threads are never removed from the thread group this caused resource
leaks when handling RPC requests. The threading has been removed as this
is handled asynchronously by ASIO anyway.

This has the added benefit of the requests being faster. When trying
getblockbynumber 25 times in a row the operation went down from 3.756s to
2.43s. A speedup of 35%.