Improve the dpor signature verification.
Record key of beacon even if it would be denied, but to diffrent variable.
Then verify that either of these keys is in boincHash. Verify the dpor
signatore against any of these keys. Accept only if any of the keys matches
and any of the signatures is valid.

Optimized alternative beacon store and relax check

fine