Use a reference time when checking if time is within a range.
Previously we used the current local time when checking beacon validity.
Since we may be way in the future when syncing but still be synced up to
beacon validity range this could cause the beacon age check to fail. This
has now been changed to validate against current best height which allows
us to sync properly.