new EstimatedTimetoStake and related functions
Corrects netweight -- replaces the old GetPosKernelPS() with a
new function GetEstimatedNetworkWeight(). (The old function does not
have anything to do with network protocol and was misleading.)

Implements new EstimatedTimetoStake() with a better algorithm and
replaces denormalized code in both GUI and RPC areas.

Implements new GetAverageDifficulty().

Changes the staking tooltip to use the correct functions.

The overview ui will be dealt with in a separate pr/commit.