Implementation of global flag fQtActive
This small pr implements a global fQtActive boolean that is set
to false in gridcoinresearch.cpp main() but true in the
main() in qt/bitcoin.cpp. This flag replaces the QT_GUI macro flag
conditionals in init.cpp and other core files to ensure that daemon
functionality is correct when a one-shot build is done of both the daemon
and GUI version of Gridcoin.

Without this change, in a one shot build the daemon will not work properly
unless --without-gui is set in configure, which then requires two separate
builds to build both the daemon and gui version.