Drop json spirit, add univalue
Univalue is a drop in replacement for json spirit. Support for json
spirit has been mostly abandoned, while univalue is actively supported by
bitcoin core developers.

Univalue is not header only, it is a full library that is compiled in
the source directory. Usage is very similar to json_spirit and was loosely based on similar
calls.