Deadlock elimination for GetEstimatedTimetoStake()
The new GetEstimatedTimetoStake() function has a rare
deadlock situation between the cs_main and cs_wallet locks taken for
a very short time in it and the `MinerStatus` locked in `GetGlobalStatus()`
and CreateCoinStake(). ifoggz suggested called GetEstimatedTimetoStake() and
assigning to a local variable right before the critical section in
GetGlobalStatus() to correct the lock order. This appears to work very
well. Wallet against testnet with the change has been running >24 hours without
deadlock.