Refactor GetJSONPollsReport
Instead of parsing everything from appcache directly into a JSON format,
we now take a step in between and add the poll contents to a poll struct.
This makes the function easier to use for calls outside the rpc context.
The polling rpc calls use a seperate function to serialize the struct to
a JSON tree.