Implementation of slightly reduced nTimeOffset
This multiplies the median of the sampled time offsets by 0.95 to take into account one's own node at 5% confidence versus rest of network at 95% confidence. It also assigns this directly to nTimeOffset (since it is not a true median anymore) and cleans up the following if statement.

This sole change will provide a gentle restoring force to return and keep the overall offset average network-wide at zero compared to the underlying consensus of system clocks. Note that since the majority of nodes are synced to UTC via some form of synchronization, such as ntp on the node, this should act to keep the network roughly in sync with UTC.