Replace broken ReadHTTP with individual calls.
The ReadHTTP function called ReadHTTPStatus without handling the return
code. This caused non-SSL servers to get stuck waiting for additional
data when it should have failed and closed the connection. The HTTP
status is for the client, not the server, so it does not make any sense
to read it there anyway.