Improve LoadAdminMessages speed.
LoadAdminMessages previously scanned 6 months back (in depth). This caused some
discrepancies in beacon acceptance since the max beacon age is 6 months in
time, not height. LoadAdminMessages was changed to scan 12 months which
circumvented the issue but also made the scan a lot slower than it needs to be.

This changes the behavior to actually scan 6 months (in time) back on full
table scans while retaining the 6 block scan on small table scans. This greatly
improves the speed of a reorganize.