Ensure sidestake address not coinstake
This is a small modification to ensure that if a sidestake is to
an output that happens to have the same address as the coinstake
address, which is entirely possible if the sidestake is local to
the staking wallet, it will be suppressed. The reward that would
be distributed will instead return in the coinstake outputs. This
makes more sense than unnecessarily separating them, and makes the
problem of displaying the right type for the stake in the UI easier.