Integration of Gridcoin C++ scraper into wallet
This is the first attempt at integrating ifoggz's scraper into
the wallet. It has been threaded from net.cpp. I have done some
tweaking to allow it to both be integrated and compiled standalone.

You can compile standalone by going to the src/ScraperProxy and
running the compile.sh command, which will generate a standalone
executable. Note that the integrated version expects the existence
of a Scraper subdirectory under the data directory to put the files,
and the standalone version expects a Scraper subdirectoy under the
CWD to put the files. I have not yet put the intelligence to create
it if it doesn't exist.

I have also changed the rpc calls in the scraper to use
local function calls when it is running as part of the wallet, since
if the scraper is running as a thread, it can skip the (loopback) rpc
and call the functions directly. I don’t like the look of the code
though and may refine it more.

Further, you may have to mess around with the environment variables

LIBS="-lcurl -ljsoncpp", because I haven't fully fixed autotools.

Tweaks to compile.sh for Scraper standalone

Added iostream to boost depends build

Changes to condition rpc calls

This makes changes to convert the rpc calls to native
function calls when the scraper is running as part of
the wallet.

It also tweaks more with some includes/defines to keep
the standalone compile working.