Implementation of filtering by beacon CPID
This is the initial pass at doing the stats by active beacons
in the 180 day window instead of by team. The beginnings of the
hashing are also in here, although the hashes are just being sent
to the log and not persisted to disk yet. I have also created a
manifest structure, but not done anything with it yet. Note that the
standalone compile is broken at the moment, because I have not
written an rpc version of GetConsensusBeaconList().