Initial implementation of manifest list
This implements a manifest list and includes the file hashes.
It is crude. It does not look for the existence of files already
in the folder, etc. Will look to improve. I also implemented a
separate GetFileHash() function for file hashing that is similar
to the one in db.cpp for peers.dat.