Scraper directory sync with mManifest
This commit puts in the logic to ensure that there is
consistency between the files in the Scraper directory
and the contents of the Manifest file. This is done at the
start of the Scraper thread.

I have also included sleep interval improvements, clamping
to the default values of the GetArg parameters, and a new parameter,
activebeforesb, which represents the number of seconds before 24
hours to wake up to begin scraping to prepare for SB formation.

So there are now three parameters...

-scraper (which defaults to false if not present, activates scraper thread)
-scrapersleep (in seconds, defaults to 60, clamped to [30, 600])
-activebeforesb (in seconds, defaults to 300, clamped to [180, 600])

To do, cull superceded project files, and implement critical section (LOCK) around
mManifest.