Implementation of stats generation
This commit contains the initial impleementation of
stats generation by the Scraper.

GetScraperStatsByConsensusBeaconList is responsible for
generating a map of network wide statistics, which contains
statistics by CPID by Project (the most granular), by Project,
by CPID, and then a final network wide entry.

It uses the helper function LoadProjectFileToStatsByCPID,
which does a single project.

I have also implemented a StoreStats function to store the map
on disk as a csv file.

This work will be useful for implementing the C++ NN.