Implement whitelist to manifest cross check
This implements a check in the scraper loop to look for and delete
manifest entries/files no longer on whitelist.

Also implement a SerializeForManifestCompare to be able to generate
a hash recorded to nContentHash at the manifest level that depends
only on content to be able to quickly compare manifests for differences.