Remove GetListOf/GetCountOf.
These functions arbitrarily filtered the results if the request was to get a
list of active beacons. These functions are now replaced with the new
GetConsensusBeaconList which has been moved to beacon.cpp/h.

In addition to this the data passed to NN::SynchronizeDPOR is now a list of
active beacons instead of an XML string containing that plus other data.
This list should be passed on to the scraper which replaces its own beacon
collecting calls.