Re-add genesis block check.
This fixes a regression introduced in 40d2a1c6d2f81f3484f2321b86f6178fd977c983
where the genesis block actually didn't pass the checks. This caused it to be
rejected and subsequent asserts to fail.

The new version is slightly more efficient as the cost is a uint256
comparison+hash check for the genesis block and just a uint256 comparison for
all other blocks.