Add excluded projects vector to StructConvergedManifest
In some instances, like we have seen from universe recently,
projects could be down past the (nominally) 48 hour retention
window. In that case the scraper convergence will have to
exclude that project, because no statistics will be available.

I have added a vExcludedProjects vector into the StructConvergedManifest
which is a vector of pairs. The pairs each contain the project
excluded and the reason.

I also changed the convergence ratio slightly to 0.60 from 2/3, because
I think 3/5 is ok, and 2/3 would require 4/5, which is too rigorous.