Scraper Authorization - Part I
This commit implements:

bool IsScraperAuthorized(), which simply wraps ALLOW_NONSCRAPER_NODE_STATS_DOWNLOAD.

bool IsScraperAuthorizedToBroadcastManifests(CKey& KeyOut).
This function checks the the wallet address, either specified by -scraperkey,
or all of the addresses in the scraper wallet, if not specified against the
authorized list of addresses in the AppCache SCRAPER section.

The function is fully implemented but neutralized for the time being, only logging
a warning rather than returning false. This is for testing purposes. As we get nearer
production, this will be changed to actually returning false.

A stub has been created in CSCraperManifest which is the complement to the above
function...
bool CScraperManifest::IsManifestAuthorized(CPubKey& PubKey).
This function currently is stubbed out to return true, but in Part II, it will
be implemented to check the PubKey of the manifest against the AppCache and error out
deserialization and increase banscore if it isn't. This will quickly isolate unauthorized
scrapers from the network.

An approach to dealing with manifest messages received while the wallet is still syncing,
where the AppCache could be not up to date and therefore the authorized list inaccurate
will also be implemented in Part II.

This receive side validation is necessary, just like ConnectBlock for blocks, because there
is no guarantee that the sending node has not been modified. So validation MUST be done on
the receiving end.