Scraper Authorization - Part II
This implements ScraperDeleteUnauthorizedCScraperManifests().

This function is called by either Scraper() or NeuralNetwork() right
after the wallet gets in sync and deletes any manifests received
while the wallet was out of sync that are not authorized. This
is necessary because the normal method of validating the received messages
against the appcache cannot be used while the wallet is out of sync,
or else the node could ban a newly authorized scraper.

This also implements bool CScraperManifest::IsManifestAuthorized(CPubKey& PubKey).

For right now it just logs an error rather than returning false during
the testing phase. This will be changed closer to production.

The only thing left is to switch ScraperSendFileManifestContents to use
the proper key in the wallet rather than the master message private key.