Make -usenewnn the default
Also default the new NN to NOT run if the scraper is specified.

Both personalities can be run simultaneously by specifying
-scraper -usenewnn. Please see code comments for details.

As part of the above, condition the thread initializers to run
either the scraper thread or the NN housekeeping thread, but not
both. It is not harmful to run both at the same time, but unnecessary.