Add team filtering functionality
Just in case we need it...

Adds network policy globals (AppCache entries)
REQUIRE_TEAM_WHITELIST_MEMBERSHIP (default to false)
TEAM_WHITELIST (default is Gridcoin, can be comma separated list of teams)

StoreTeamIDList
LoadTeamIDList
DownloadProjectTeamFiles
ProcessProjectTeamFiles

and makes modifications to several areas, including ProcessProjectRacFileByCPID,
to accomodate team filtering.

Note that team ID map is stored as an in memory file and also persisted to disk,
and team files are not downloaded once the map is populated, unless a project is added
or subtracted, or the team whitelist is changed.