fix seg fault part II - check for existence in mapManifest
It is possible, because the housekeeping thread periodically
checks for unauthorized and/or old manifests and deletes the map
entries for those, that in the split second between the passing in
of the inv object and the actual check of the map to send the manifest,
it could be deleted. The fix is to simply check and see if it is still there
before the assignment.