Adjust GetmScraperFileManifestHash
Every once in a while CScraperManifest hash sends would fail. These
failures were innocuous, but irritating. I traced it down to needing a
filter condition on the map hash calculated in GetmScraperFileManifestHash.

The purpose of this hash is primarily to allow the scraper to detect when
to publish a manifest when a new file is pulled down from a BOINC site.

Before, the function was hashing all file entries in the map, whether they
were current or not. This is incorrect, as the hash set should actually be
conditioned on the boolean configuration option SCRAPER_CMANIFEST_INCLUDE_NONCURRENT_PROJ_FILES.

If SCRAPER_CMANIFEST_INCLUDE_NONCURRENT_PROJ_FILES is true, then all files
should be included; however, if it is false, only the CURRENT files in the
map should be included, because the manifest published will only include
current files.

Due to the lack of filtering before, if a non-current manifest was deleted
(from the map) after a CSCraperManifest was already published, due to aging rules,
but before a new file is downloaded during the download process. The scraper would see
the deletion of the old file as a change and attempt to publish a new manifest. This
is correct behavior if the manifest is to include ALL files, but incorrect if
the manifest only is to include current files. Hence the if condition for the filter.