scraper.log archiving
This commit implements a simple log archiver for the scraper log.

If people like this I will do something similar for the main log
as well.

Currently it runs in the housekeeping loop and automatically
renames the scraper.log to scraper-YYYYMMDDHHMMSS.log then
compresses the archive log to a gz file. A lock is held only during
the rename, and a new scraper.log is started immediately.

Stream compression (just like the rest of the scraper) is used to
minimize memory consumption.

I have also provided an rpc command, archivescraperlog, which takes
no arguments and can be used to immediately roll the current log
to an archive file and start a new one non-disruptively.