Add stake splitting and side staking info to getmininginfo
I refactored the parameter parsing in miner.cpp into separate
callable functions that can be used safely from both the miner and
rpc calls.

getmininginfo has been changed to include information on the status
and parameters of stake splitting and side staking. It also includes
the UTXO size needed to stake at the supplied efficiency in the
config file.