Initial roughout of consolidateutxos
Includes a new overloaded CreateTransaction function.

This function works by rpc call...

consolidateutxos <address>
            [utxo size limit [max num of inputs]]

The address is an address in your wallet for which you
want to combine UTXOs. The optional utxo size limit means
to try and make an output of that size.

The function consolidates UTXO's starting from the smallest
ascending order on that address, stopping when either the total
output size exceeds the utxo size limit (if specified) OR max num
of inputs (selected UTXO's) have been reached OR all UTXO's on that
address have been selected, whichever occurs first. The max num
of inputs is clamped to 200.

The fees are calculated and netted to produce no change, but
coincontrol is used to route the change back to the same address
just in case.