Add support for explorer mode
This adds support for the -explorer flag, which changes the
behavior of the scraper to hold files for a longer period of time
and also download team and host files. The publishing of manifests
is not affected.

This is the initial implementation of the explorer flag, team and host
file downloading and retention.