Improve convergence and SB contract caching.
This implements a bClean boolean in that is marked false
in scraper_net when manifests are received from the network
or published locally. It is marked true when a new set of
statistics and SBContract core is computed.

The rule is that the cached contract will be used when the cache age
is younger than nScraperSleep in seconds OR the cache is clean (i.e.
no new manifests have been published (if a scraper node) or received
(if a normal node). This will avoid the statistics
calculation pulse seen on mainnet every 300 seconds during times
the scrapers are not active and publishing new manifests.