Move legacy neural vote warnings to debug log level
This suppresses log messages that indicate that a block contains an
invalid superblock vote. A large span of old blocks contain invalid
votes which result in thousands of messages. Sync speed may improve
a bit by showing them in the debug log level only. The messages are
not particularly actionable, so most users don't need to see these.