Remove useless "tally" and "tallyneural" RPC methods
The "tally" RPC method triggers a count of the two-week network averages
for research payment calculation. The height for a tally recount is very
specific, so ths RPC method just counts from the last tally block height
and resets the averages. The wallet does this anyway when blocks arrive,
so this RPC call doesn't actually change any data.

The "tallyneural" RPC call forces a quorum vote recount. The application
already does this before generating and after receiving a superblock, so
there's no need to re-count the votes manually--the next quorum interval
will overwrite the data anyway.